Allowing for connections to local host
The local host cannot be connected to with a username and password.
Previous code didn't account for this and so the connection failed.

Now wrapped connection code in a try...catch block.  If the initial
connection fails with an error code indicating a localhost error, then
the connection is attempted again with a null username and password.